pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "sti-munene",email = "kinyuamunene.km@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-ckeditor-5 (>=0.2.20,<0.3.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "faker (>=40.15.0,<41.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django-tree-queries (>=0.24.0,<0.25.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
